Refactor module to add support for RedHat (and SUSE).
This is a fairly extensive refactoring to make it easier to add
support for additional repos.  It'll continue to use the same
repo for Debian/Ubuntu but introduces a different repo for RHEL5.

We use the stock (hopefully EPEL) erlang for Redhat, and the official
SUSE repo for erlang on that side.

In addition this fixes up the apt support on the Debian side to fix
a few bugs I found.