Add EPEL support.
While testing centos5/6 I realized it was going to fail as we require
EPEL to be enabled and the testing images don't have it.  Add
`epel_enable` as an option in case people are already handling epel
in another way.

This fixes the centos6 case which didn't work out of the box.